Bug #2425024 unisoc_platform: sched: fix int and unsigned long
[root cause  ]: fix int add unsigned long
[changes     ]: fix int add unsigned long
[side effects]: no
[self test   ]: yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:yes
[board]:n6l
[test case]:camera
[reviewers]:ke

commit_template_version:v1

int add unsigned long, int will be transformed to unsigned long.

+++ drivers/unisoc_platform/sched/uni_sched.h
@@ -389,9 +389,9 @@ static inline int group_boost_margin(long util, int boost)
 static inline unsigned long boosted_task_util(struct task_struct *task)
 {
 #if IS_ENABLED(CONFIG_SCHED_WALT)
-	unsigned long util = walt_task_util(task);
+	int util = walt_task_util(task);
 #else
-	unsigned long util = task_util_est(task);
+	int util = task_util_est(task);
 #endif
 	int boost = task_group_boost(task);
 	int margin = group_boost_margin(util, boost);
